+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 New features:
 
+* Add a lightweight callback-based producer metrics collector protocol
+  (issue #1166)
 * Change the way tagged fields are managed in the protocol definitions
   (pr #1162 by @vmaurin)